rename redirection-page file
